Section - 2/requirements.txt
langchain==0.1.0
langchain-community==0.0.10
langchain-core==0.1.10
langchain-openai==0.0.3
openai==1.6.1
streamlit==1.29.0
python-dotenv==1.0.0
psycopg2-binary==2.9.9
pandas==2.1.4
pypdf==3.17.1
python-magic==0.4.27
tiktoken==0.5.2
faiss-cpu==1.7.4
numpy==1.26.2
pydantic==2.5.3
chromadb==0.4.18